pyproject.toml
[project]
name = "supercell"
version = "1.0.0"
description = "SUPERCELL — AI-powered ATC emergency triage environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "httpx>=0.28.0",
    "openai>=1.60.0",
    "openenv-core>=0.2.0",
    "matplotlib>=3.10.8",
    "fpdf2>=2.8.7",
    "pillow>=12.1.1",
    "edge-tts>=7.2.8",
    "transformers>=4.44.0",
    "huggingface_hub>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps/api/server", "apps/api/models.py", "apps/api/client.py"]

[tool.pytest.ini_options]
testpaths = ["apps/api/tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.optional-dependencies]
local = [
    "torch>=2.1.0",
]
